+++ src/main/java/com/buuz135/simpleclaims/gui/PartyInfoEditGui.java
@@ -2,9 +2,11 @@ package com.buuz135.simpleclaims.gui;
 
 import com.buuz135.simpleclaims.claim.ClaimManager;
 import com.buuz135.simpleclaims.claim.party.PartyInfo;
+import com.buuz135.simpleclaims.claim.party.PartyInvite;
 import com.buuz135.simpleclaims.claim.party.PartyOverride;
 import com.buuz135.simpleclaims.claim.party.PartyOverrides;
 import com.buuz135.simpleclaims.Main;
+import com.buuz135.simpleclaims.commands.CommandMessages;
 import com.hypixel.hytale.codec.Codec;
 import com.hypixel.hytale.codec.KeyedCodec;
 import com.hypixel.hytale.codec.builder.BuilderCodec;
@@ -13,13 +15,20 @@ import com.hypixel.hytale.component.Store;
 import com.hypixel.hytale.protocol.packets.interface_.CustomPageLifetime;
 import com.hypixel.hytale.protocol.packets.interface_.CustomUIEventBindingType;
 import com.hypixel.hytale.server.core.entity.entities.player.pages.InteractiveCustomUIPage;
+import com.hypixel.hytale.server.core.ui.DropdownEntryInfo;
+import com.hypixel.hytale.server.core.ui.LocalizableString;
 import com.hypixel.hytale.server.core.ui.builder.EventData;
 import com.hypixel.hytale.server.core.ui.builder.UICommandBuilder;
 import com.hypixel.hytale.server.core.ui.builder.UIEventBuilder;
 import com.hypixel.hytale.server.core.universe.PlayerRef;
+import com.hypixel.hytale.server.core.universe.Universe;
 import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
 import com.hypixel.hytale.server.worldgen.loader.util.ColorUtil;
 import org.checkerframework.checker.nullness.compatqual.NonNullDecl;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
 
 public class PartyInfoEditGui extends InteractiveCustomUIPage<PartyInfoEditGui.PartyInfoData> {
 
@@ -28,6 +37,8 @@ public class PartyInfoEditGui extends InteractiveCustomUIPage<PartyInfoEditGui.P
     private String description;
     private int requestingConfirmation;
     private final boolean isOpEdit;
+    private String inviteDropdown;
+    private String alliesDropdown;
 
     public PartyInfoEditGui(@NonNullDecl PlayerRef playerRef, PartyInfo info, boolean isOpEdit) {
         super(playerRef, CustomPageLifetime.CanDismiss, PartyInfoData.CODEC);
@@ -68,6 +79,33 @@ public class PartyInfoEditGui extends InteractiveCustomUIPage<PartyInfoEditGui.P
         if (data.removeButtonAction != null) {
             var split = data.removeButtonAction.split(":");
             var action = split[0];
+            if (action.equals("DeleteInvite")) {
+                ClaimManager.getInstance().getPartyInvites().remove(UUID.fromString(split[1]));
+                ClaimManager.getInstance().markDirty();
+                UICommandBuilder commandBuilder = new UICommandBuilder();
+                UIEventBuilder eventBuilder = new UIEventBuilder();
+                this.build(ref, commandBuilder, eventBuilder, store);
+                this.sendUpdate(commandBuilder, eventBuilder, true);
+                return;
+            }
+            if (action.equals("DeleteAllyPlayer")) {
+                this.info.getPlayerAllies().remove(UUID.fromString(split[1]));
+                ClaimManager.getInstance().markDirty();
+                UICommandBuilder commandBuilder = new UICommandBuilder();
+                UIEventBuilder eventBuilder = new UIEventBuilder();
+                this.build(ref, commandBuilder, eventBuilder, store);
+                this.sendUpdate(commandBuilder, eventBuilder, true);
+                return;
+            }
+            if (action.equals("DeleteAllyParty")) {
+                this.info.getPartyAllies().remove(UUID.fromString(split[1]));
+                ClaimManager.getInstance().markDirty();
+                UICommandBuilder commandBuilder = new UICommandBuilder();
+                UIEventBuilder eventBuilder = new UIEventBuilder();
+                this.build(ref, commandBuilder, eventBuilder, store);
+                this.sendUpdate(commandBuilder, eventBuilder, true);
+                return;
+            }
             var index = Integer.parseInt(split[1]);
             if (action.equals("Click")){
                 this.requestingConfirmation = index;
@@ -101,6 +139,52 @@ public class PartyInfoEditGui extends InteractiveCustomUIPage<PartyInfoEditGui.P
                 System.out.println("Invalid color");
             }
         }
+        if (data.inviteDropdown != null) {
+            this.inviteDropdown = data.inviteDropdown;
+        }
+        if (data.alliesDropdown != null) {
+            this.alliesDropdown = data.alliesDropdown;
+        }
+        if (data.button != null) {
+            if (data.button.equals("Invite") && this.inviteDropdown != null) {
+                if (!this.info.isMember(UUID.fromString(this.inviteDropdown))) {
+                    var invited = Universe.get().getPlayer(UUID.fromString(this.inviteDropdown));
+                    if (invited != null) {
+                        ClaimManager.getInstance().invitePlayerToParty(this.playerRef, this.info, invited);
+                        ClaimManager.getInstance().markDirty();
+                        invited.sendMessage(CommandMessages.PARTY_INVITE_RECEIVED.param("party_name", this.info.getName()).param("username", this.playerRef.getUsername()));
+                        UICommandBuilder commandBuilder = new UICommandBuilder();
+                        UIEventBuilder eventBuilder = new UIEventBuilder();
+                        this.build(ref, commandBuilder, eventBuilder, store);
+                        this.sendUpdate(commandBuilder, eventBuilder, true);
+                        return;
+                    }
+                }
+            }
+            if (data.button.equals("Allies") && this.alliesDropdown != null) {
+                var invited = Universe.get().getPlayer(UUID.fromString(this.alliesDropdown));
+                if (invited != null) { //IS Player
+                    this.info.getPlayerAllies().add(invited.getUuid());
+                    ClaimManager.getInstance().markDirty();
+                    UICommandBuilder commandBuilder = new UICommandBuilder();
+                    UIEventBuilder eventBuilder = new UIEventBuilder();
+                    this.build(ref, commandBuilder, eventBuilder, store);
+                    this.sendUpdate(commandBuilder, eventBuilder, true);
+                    return;
+                } else {
+                    var party = ClaimManager.getInstance().getPartyById(UUID.fromString(this.alliesDropdown));
+                    if (party != null) {
+                        this.info.getPartyAllies().add(party.getId());
+                        ClaimManager.getInstance().markDirty();
+                        UICommandBuilder commandBuilder = new UICommandBuilder();
+                        UIEventBuilder eventBuilder = new UIEventBuilder();
+                        this.build(ref, commandBuilder, eventBuilder, store);
+                        this.sendUpdate(commandBuilder, eventBuilder, true);
+                        return;
+                    }
+                }
+            }
+        }
         this.sendUpdate();
     }
 
@@ -115,7 +199,8 @@ public class PartyInfoEditGui extends InteractiveCustomUIPage<PartyInfoEditGui.P
         uiCommandBuilder.set("#PartyInfo #PartyDescriptionField.Value", this.info.getDescription());
         uiCommandBuilder.set("#PartyInfo #PartyDescriptionField.IsReadOnly", !playerCanModify);
         uiEventBuilder.addEventBinding(CustomUIEventBindingType.ValueChanged, "#PartyDescriptionField", EventData.of("@Description", "#PartyDescriptionField.Value"), false);
-        for (int i = 0; i < this.info.getMembers().length; i++) {
+        int i = 0;
+        for (; i < this.info.getMembers().length; i++) {
             uiCommandBuilder.append("#MemberEntries", "Pages/Buuz135_SimpleClaims_PartyMemberListEntry.ui");
             var isOwner = this.info.isOwner(this.info.getMembers()[i]);
             uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberName.Text", ClaimManager.getInstance().getPlayerNameTracker().getPlayerName(this.info.getMembers()[i]));
@@ -137,6 +222,82 @@ public class PartyInfoEditGui extends InteractiveCustomUIPage<PartyInfoEditGui.P
                 }
             }
         }
+        for (PartyInvite value : ClaimManager.getInstance().getPartyInvites().values()) {
+            uiCommandBuilder.append("#MemberEntries", "Pages/Buuz135_SimpleClaims_PartyMemberListEntry.ui");
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberName.Text", ClaimManager.getInstance().getPlayerNameTracker().getPlayerName(value.recipient()));
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberRole.Text", "Pending Invite");
+
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberRole.Background.Color", "#cac85383");
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberRole.OutlineColor", "#cac853de");
+
+            if (!playerCanModify) {
+                uiCommandBuilder.set("#MemberEntries[" + i + "] #RemoveMemberButton.Disabled", true);
+            } else {
+                //uiEventBuilder.addEventBinding(CustomUIEventBindingType.SlotMouseExited, "#Members[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "Left:" + i), false);
+                if (this.requestingConfirmation == i) {
+                    uiCommandBuilder.set("#MemberEntries[" + i + "] #RemoveMemberButton.Text", "Are you sure?");
+                    uiEventBuilder.addEventBinding(CustomUIEventBindingType.Activating, "#MemberEntries[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "DeleteInvite:" + value.recipient().toString()), false);
+                    uiEventBuilder.addEventBinding(CustomUIEventBindingType.MouseExited, "#MemberEntries[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "Click:-1"), false);
+                } else {
+                    uiEventBuilder.addEventBinding(CustomUIEventBindingType.Activating, "#MemberEntries[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "Click:" + i), false);
+                }
+            }
+
+            ++i;
+        }
+
+        for (UUID uuid : this.info.getPlayerAllies()) {
+            uiCommandBuilder.append("#MemberEntries", "Pages/Buuz135_SimpleClaims_PartyMemberListEntry.ui");
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberName.Text", ClaimManager.getInstance().getPlayerNameTracker().getPlayerName(uuid));
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberRole.Text", "  Ally  ");
+
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberRole.Background.Color", "#5ab44e83");
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberRole.OutlineColor", "#5ab44ede");
+
+            if (!playerCanModify) {
+                uiCommandBuilder.set("#MemberEntries[" + i + "] #RemoveMemberButton.Disabled", true);
+            } else {
+                //uiEventBuilder.addEventBinding(CustomUIEventBindingType.SlotMouseExited, "#Members[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "Left:" + i), false);
+                if (this.requestingConfirmation == i) {
+                    uiCommandBuilder.set("#MemberEntries[" + i + "] #RemoveMemberButton.Text", "Are you sure?");
+                    uiEventBuilder.addEventBinding(CustomUIEventBindingType.Activating, "#MemberEntries[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "DeleteAllyPlayer:" + uuid.toString()), false);
+                    uiEventBuilder.addEventBinding(CustomUIEventBindingType.MouseExited, "#MemberEntries[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "Click:-1"), false);
+                } else {
+                    uiEventBuilder.addEventBinding(CustomUIEventBindingType.Activating, "#MemberEntries[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "Click:" + i), false);
+                }
+            }
+
+            ++i;
+        }
+
+        for (UUID uuid : this.info.getPartyAllies()) {
+            uiCommandBuilder.append("#MemberEntries", "Pages/Buuz135_SimpleClaims_PartyMemberListEntry.ui");
+            var name = "Unknown Party";
+            if (ClaimManager.getInstance().getParties().containsKey(uuid.toString())) {
+                name = ClaimManager.getInstance().getParties().get(uuid.toString()).getName();
+            }
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberName.Text", name);
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberRole.Text", "Party Ally");
+
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberRole.Background.Color", "#5ab44e83");
+            uiCommandBuilder.set("#MemberEntries[" + i + "] #MemberRole.OutlineColor", "#5ab44ede");
+
+            if (!playerCanModify) {
+                uiCommandBuilder.set("#MemberEntries[" + i + "] #RemoveMemberButton.Disabled", true);
+            } else {
+                //uiEventBuilder.addEventBinding(CustomUIEventBindingType.SlotMouseExited, "#Members[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "Left:" + i), false);
+                if (this.requestingConfirmation == i) {
+                    uiCommandBuilder.set("#MemberEntries[" + i + "] #RemoveMemberButton.Text", "Are you sure?");
+                    uiEventBuilder.addEventBinding(CustomUIEventBindingType.Activating, "#MemberEntries[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "DeleteAllyParty:" + uuid.toString()), false);
+                    uiEventBuilder.addEventBinding(CustomUIEventBindingType.MouseExited, "#MemberEntries[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "Click:-1"), false);
+                } else {
+                    uiEventBuilder.addEventBinding(CustomUIEventBindingType.Activating, "#MemberEntries[" + i + "] #RemoveMemberButton", EventData.of("RemoveButtonAction", "Click:" + i), false);
+                }
+            }
+
+            ++i;
+        }
+
         uiEventBuilder.addEventBinding(CustomUIEventBindingType.Activating, "#SaveChangesButton", EventData.of("Save", "true"), false);
         uiEventBuilder.addEventBinding(CustomUIEventBindingType.Activating, "#CancelButton", EventData.of("Cancel", "true"), false);
 
@@ -162,6 +323,21 @@ public class PartyInfoEditGui extends InteractiveCustomUIPage<PartyInfoEditGui.P
         uiCommandBuilder.set("#ClaimedChunksInfo #MaxChunksCount.Text",this.info.getMaxClaimAmount() + "");
 
         uiEventBuilder.addEventBinding(CustomUIEventBindingType.ValueChanged, "#ClaimColorPicker", EventData.of("@ClaimColor", "#ClaimColorPicker.Value"), false);
+
+        //Invite Dropdowns
+        var players = new ArrayList<>(Universe.get().getPlayers().stream().map(playerRef1 -> new DropdownEntryInfo(LocalizableString.fromString(playerRef1.getUsername()), playerRef1.getUuid().toString())).toList());
+        uiCommandBuilder.set("#InviteDropdown.Entries", players);
+        var parties = ClaimManager.getInstance().getParties().values().stream().map(party -> new DropdownEntryInfo(LocalizableString.fromString("[PAR] " + party.getName()), party.getId().toString())).toList();
+        players.addAll(parties);
+        uiCommandBuilder.set("#AlliesDropdown.Entries", players);
+
+        uiEventBuilder.addEventBinding(CustomUIEventBindingType.ValueChanged, "#InviteDropdown", EventData.of("@InviteDropdown", "#InviteDropdown.Value"), false);
+        uiEventBuilder.addEventBinding(CustomUIEventBindingType.ValueChanged, "#AlliesDropdown", EventData.of("@AlliesDropdown", "#AlliesDropdown.Value"), false);
+        uiEventBuilder.addEventBinding(CustomUIEventBindingType.Activating, "#ConfirmInviteButton", EventData.of("Button", "Invite"), false);
+        uiEventBuilder.addEventBinding(CustomUIEventBindingType.Activating, "#ConfirmAlliesButton", EventData.of("Button", "Allies"), false);
+
+        uiCommandBuilder.set("#ConfirmInviteButton.Disabled", !playerCanModify);
+        uiCommandBuilder.set("#ConfirmAlliesButton.Disabled", !playerCanModify);
     }
 
     public static class PartyInfoData {
@@ -171,6 +347,10 @@ public class PartyInfoEditGui extends InteractiveCustomUIPage<PartyInfoEditGui.P
         static final String KEY_CANCEL = "Cancel";
         static final String KEY_REMOVE_BUTTON_ACTION = "RemoveButtonAction";
         static final String KEY_CLAIM_COLOR = "@ClaimColor";
+        static final String KEY_INVITE_DROPDOWN = "@InviteDropdown";
+        static final String KEY_ALLIES_DROPDOWN = "@AlliesDropdown";
+        static final String KEY_BUTTON = "Button";
+
         public static final BuilderCodec<PartyInfoData> CODEC = BuilderCodec.<PartyInfoData>builder(PartyInfoData.class, PartyInfoData::new)
                 .addField(new KeyedCodec<>(KEY_NAME, Codec.STRING), (searchGuiData, s) -> searchGuiData.name = s, searchGuiData -> searchGuiData.name)
                 .addField(new KeyedCodec<>(KEY_DESCRIPTION, Codec.STRING), (searchGuiData, s) -> searchGuiData.description = s, searchGuiData -> searchGuiData.description)
@@ -178,7 +358,9 @@ public class PartyInfoEditGui extends InteractiveCustomUIPage<PartyInfoEditGui.P
                 .addField(new KeyedCodec<>(KEY_CANCEL, Codec.STRING), (searchGuiData, s) -> searchGuiData.cancel = s, searchGuiData -> searchGuiData.cancel)
                 .addField(new KeyedCodec<>(KEY_REMOVE_BUTTON_ACTION, Codec.STRING), (searchGuiData, s) -> searchGuiData.removeButtonAction = s, searchGuiData -> searchGuiData.removeButtonAction)
                 .addField(new KeyedCodec<>(KEY_CLAIM_COLOR, Codec.STRING), (searchGuiData, s) -> searchGuiData.claimColor = s, searchGuiData -> searchGuiData.claimColor)
-
+                .addField(new KeyedCodec<>(KEY_INVITE_DROPDOWN, Codec.STRING), (searchGuiData, s) -> searchGuiData.inviteDropdown = s, searchGuiData -> searchGuiData.inviteDropdown)
+                .addField(new KeyedCodec<>(KEY_ALLIES_DROPDOWN, Codec.STRING), (searchGuiData, s) -> searchGuiData.alliesDropdown = s, searchGuiData -> searchGuiData.alliesDropdown)
+                .addField(new KeyedCodec<>(KEY_BUTTON, Codec.STRING), (searchGuiData, s) -> searchGuiData.button = s, searchGuiData -> searchGuiData.button)
                 .build();
 
         private String name;
@@ -187,6 +369,9 @@ public class PartyInfoEditGui extends InteractiveCustomUIPage<PartyInfoEditGui.P
         private String cancel;
         private String removeButtonAction;
         private String claimColor;
+        private String inviteDropdown;
+        private String alliesDropdown;
+        private String button;
 
         @Override
         public String toString() {